Store jiraUpdates in react state
