+++ src/app/state/JiraUpdatesState.ts
@@ -1,0 +1,9 @@
+import { atom } from 'recoil';
+import { JiraUpdate } from '../../shared/dto/JiraTypes';
+
+const jiraUpdatesState = atom<JiraUpdate[]>({
+  key: 'jiraUpdate',
+  default: [],
+});
+
+export default jiraUpdatesState;
